[ADD] pacgum effect to kill ghost

+++ lib/arcade_pacman/src/class/GamePacman.cpp
@@ -115,16 +115,24 @@ void GamePacman::update(std::size_t deltaTime) {
         this->openDoor();
     std::size_t index = 0;
     for (std::vector<pacman::Enemy>::iterator i = this->listEnemies.begin(); i != this->listEnemies.end(); i++) {
+        i->waiting(deltaTime);
         i->vulnerable(deltaTime < this->endEffect);
         Input direction = QUIT;
         if (deltaTime > 10000 * (index + 1))
             direction = i->chooseDirection(this->player, this->map);
-        i->move(direction, this->map);
-        if (std::round(this->player.getPosition().x) == i->getPosition().x && std::round(this->player.getPosition().y) == i->getPosition().y) {
-            this->player.kill();
-            this->lives--;
-            if (this->lives >= 0)
-                this->listLives[this->lives].setVisibility(false);
+        if (i->getEntityType() != ENEMY5_DYING)
+            i->move(direction, this->map);
+        if (std::round(this->player.getPosition().x) == i->getPosition().x && std::round(this->player.getPosition().y) == i->getPosition().y && i->getEntityType() != ENEMY5_DYING) {
+            if (i->getEntityType() == ENEMY5) {
+                this->score += 1000;
+                *i = pacman::Enemy(Vector2D(10, 10, RIGHT), i->getType());
+                i->kill(deltaTime);
+            } else {
+                this->player.kill();
+                this->lives--;
+                if (this->lives >= 0)
+                    this->listLives[this->lives].setVisibility(false);
+            }
         }
         index++;
     }
@@ -225,7 +233,8 @@ std::map<EntityType, std::pair<std::string, std::size_t>> GamePacman::getSpriteD
         {EntityType::ENEMY2, {std::string("lib/assets/pacman_enemy2.png"), 2}},
         {EntityType::ENEMY3, {std::string("lib/assets/pacman_enemy3.png"), 2}},
         {EntityType::ENEMY4, {std::string("lib/assets/pacman_enemy4.png"), 2}},
-        {EntityType::ENEMY5, {std::string("lib/assets/pacman_enemy5.png"), 2}}
+        {EntityType::ENEMY5, {std::string("lib/assets/pacman_enemy5.png"), 2}},
+        {EntityType::ENEMY5_DYING, {std::string("lib/assets/pacman_enemy5_dying.png"), 2}}
     };
     return dict;
 }
